:art: Refactor padding in search form and improve header logo visibility based on screen size :lipstick: Add icon styling for GodPraksisHeroButton and create GodPraksisTaxonomyTag component
