+++ aksel.nav.no/website/app/_ui/global-search/GlobalSearch.module.css
@@ -3,7 +3,7 @@
   display: flex;
   align-items: center;
   gap: var(--ax-space-8);
-  padding: var(--ax-space-4) var(--ax-space-8);
+  padding: var(--ax-space-8);
   border-bottom: 1px solid var(--ax-border-neutral-subtle);
   background-color: var(--ax-bg-neutral-soft);
