+++ aksel.nav.no/website/app/dev/(god-praksis)/_ui/GodPraksisTaxonomyTag.tsx
@@ -1,0 +1,32 @@
+import { FileFillIcon, TagFillIcon } from "@navikt/aksel-icons";
+import { Tag } from "@navikt/ds-react";
+
+function GodPraksisTaxonomyTag({
+  children,
+  type,
+}: {
+  children: React.ReactNode;
+  type: "innholdstype" | "undertema";
+}) {
+  if (!children) {
+    return null;
+  }
+
+  return (
+    <Tag
+      variant={type === "undertema" ? "alt3-moderate" : "alt1-moderate"}
+      size="xsmall"
+      icon={
+        type === "undertema" ? (
+          <TagFillIcon aria-hidden />
+        ) : (
+          <FileFillIcon aria-hidden />
+        )
+      }
+    >
+      {children}
+    </Tag>
+  );
+}
+
+export { GodPraksisTaxonomyTag };
