+++ aksel.nav.no/website/app/dev/(god-praksis)/_ui/hero/Hero.module.css
@@ -107,3 +107,7 @@
     visibility: hidden;
   }
 }
+
+.godPraksisHeroButtonIcon {
+  flex-shrink: 0;
+}
